Move Windows util code into test sources
