+++ core/src/test/java/hudson/RemoveWindowsDirectoryJunctionTest.java
@@ -8,6 +8,8 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assume.assumeTrue;
 
 import java.io.File;
+
+import hudson.os.WindowsUtil;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -32,7 +34,7 @@ public class RemoveWindowsDirectoryJunctionTest {
         File subdir1 = tmp.newFolder("notJunction");
         File f1 = new File(subdir1, "testfile1.txt");
         assertTrue("Unable to create temporary file in notJunction directory", f1.createNewFile());
-        File j1 = Util.createJunction(tmp.getRoot(), subdir1);
+        File j1 = WindowsUtil.createJunction(tmp.getRoot(), subdir1);
         Util.deleteRecursive(j1);
         assertFalse("Windows Junction should have been removed", j1.exists());
         assertTrue("Contents of Windows Junction should not be removed", f1.exists());
